Craigslist used Car Crawler
Now crawls craigslist for keyword search car and exports titles and price to CSV. Next steps is fixing an issue of price being duplicated that needs to be fixed. As well as adding the correct link to the posts to the csv file as well. Also to filter by price based on how much a user wants to pay